moving test specific fixture to test file
